requirements.txt
# Web Framework
flask==2.2.3
Werkzeug==2.2.3
gunicorn==23.0.0
flask-cors==3.0.10
jinja2==3.1.2

# Environment Variables
python-dotenv==1.0.0

# API Clients
requests==2.31.0
httpx==0.23.3

# Data Processing
python-dateutil==2.8.2

# File Processing
python-docx==0.8.11
PyPDF2==3.0.1
docx2txt==0.8

# Security & Networking
urllib3==2.0.7

# System Monitoring
psutil==5.9.5

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Database
supabase==2.4.0

# OpenAI
openai==1.14.0

# PDF Generation
pdflatex==0.1.3
Pillow==11.2.1

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Document Processing
python-pptx==0.6.21    # PowerPoint file parsing
pdfminer.six==20221105 # Advanced PDF text extraction

# LaTeX Generation
pdflatex==0.1.3        # Python wrapper for pdflatex (optional, system pdflatex preferred)

# Utility Libraries
tqdm==4.66.1           # Progress bars
pyyaml==6.0.1          # YAML parsing and generation
charset-normalizer==3.3.2  # Character encoding detection
idna==3.6              # Internationalized Domain Names
certifi==2023.11.17    # Root certificates for validation
typing-extensions==4.9.0  # Typing extensions for type hints 
 